requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
absl-py==0.7.0            # via tensorflow
astor==0.7.1              # via tensorflow
click==7.0                # via pip-tools
cycler==0.10.0            # via matplotlib
gast==0.2.2               # via tensorflow
grpcio==1.18.0            # via tensorboard, tensorflow
h5py==2.9.0               # via keras-applications
keras-applications==1.0.6  # via tensorflow
keras-preprocessing==1.0.5  # via tensorflow
kiwisolver==1.0.1         # via matplotlib
markdown==3.0.1           # via tensorboard
matplotlib==3.0.2
numpy==1.16.0             # via h5py, keras-applications, keras-preprocessing, matplotlib, pandas, scikit-learn, scipy, tensorboard, tensorflow
pandas==0.23.4
pip-tools==3.2.0
protobuf==3.6.1           # via tensorboard, tensorflow
pyparsing==2.3.1          # via matplotlib
python-dateutil==2.7.5    # via matplotlib, pandas
pytz==2018.9              # via pandas
scikit-learn==0.20.2
scipy==1.2.0              # via scikit-learn
six==1.12.0               # via absl-py, cycler, grpcio, h5py, keras-preprocessing, pip-tools, protobuf, python-dateutil, tensorboard, tensorflow
tensorboard==1.12.2       # via tensorflow
tensorflow==1.12.0
termcolor==1.1.0          # via tensorflow
werkzeug==0.14.1          # via tensorboard
wheel==0.32.3             # via tensorboard, tensorflow
